Split up TUI code
